requirements.txt
azure-functions
azure-functions-worker
azure-storage-blob
azure-storage
pyodbc
pypyodbc
numpy
imagehash
matplotlib
scipy
sklearn